AppleDeviceSource: Fix items disappearing in playlists on re-connect
We store GPod.Track.DBID in the ExternalID field, which is handled as a
long, so we need an explicit cast. Fixes bgo#650011.

+++ src/Dap/Banshee.Dap.AppleDevice/Banshee.Dap.AppleDevice/AppleDeviceSource.cs
@@ -238,10 +238,11 @@ namespace Banshee.Dap.AppleDevice
 
                 PlaylistSource pl_src = new PlaylistSource (playlist.Name, this);
                 pl_src.Save ();
-                // We use the IPod.Track.Id here b/c we just shoved it into ExternalID above when we loaded
-                // the tracks, however when we sync, the Track.Id values may/will change.
+                // We use the GPod.Track.DBID here b/c we just shoved it into ExternalID above when we loaded
+                // the tracks, however when we sync, the Track.DBID values may/will change.
                 foreach (var track in playlist.Tracks) {
-                    ServiceManager.DbConnection.Execute (insert_cmd, pl_src.DbId, this.DbId, track.DBID);
+                    // DBID will be stored in a long, so we need to cast it. See bgo#650011
+                    ServiceManager.DbConnection.Execute (insert_cmd, pl_src.DbId, this.DbId, (long) track.DBID);
                 }
                 pl_src.UpdateCounts ();
                 AddChildSource (pl_src);
